Moved logic from EditPost.tsx into new useEditPost hook.
